\name{CItrans}
\alias{exp.mcpCI}
\alias{expit.mcpCI}

\title{Transformation of Confidence Intervals}
\description{Transformation of confidence interval estimates in mcpCI objects.}
\usage{
\method{exp}{mcpCI}(x)
\method{expit}{mcpCI}(x)
}

\arguments{
  \item{x}{An object of class mcpCI}
}

\value{An object of class mcpCI with transformed estimates.}

\author{Daniel Gerhard}

\seealso{\code{\link{exp}}, \code{\link{confint.mcprofile}}}

\keyword{misc}
