% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cta}
\alias{cta}
\title{Cell transformation assay dataset}
\format{A data frame with 80 observations on the following 2 variables.
\describe{
  \item{\code{conc}}{a numeric vector denoting the concentration levels}
  \item{\code{foci}}{a numeric vector with the number of foci}
}}
\usage{
cta
}
\description{
Balb//c 3T3 cells are treated with different concentrations of a carcinogen. Cells treated with a carcinogen do not stop proliferation. Number of foci (cell accumulations) are counted for 10 replicates per concentration level.
}
\references{
Thomas C (2008): ECVAM data
}
\keyword{datasets}
