% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CItrans.R
\name{exp.mcpCI}
\alias{exp.mcpCI}
\title{Transformation of Confidence Intervals}
\usage{
\method{exp}{mcpCI}(x)
}
\arguments{
\item{x}{An object of class mcpCI}
}
\value{
An object of class mcpCI with transformed estimates.
}
\description{
Transformation of confidence interval estimates in mcpCI objects.
}
\seealso{
\code{\link{exp}}, \code{\link{confint.mcprofile}}
}
\keyword{misc}
