% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.mcpCI.R
\name{expit.mcpCI}
\alias{expit.mcpCI}
\title{Inverse logit transformation of Confidence Intervals}
\usage{
expit.mcpCI(x)
}
\arguments{
\item{x}{An object of class mcpCI}
}
\value{
An object of class mcpCI with transformed estimates.
}
\description{
Inverse logit transformation of confidence interval estimates in mcpCI objects.
}
\seealso{
\code{\link{exp}}, \code{\link{confint.mcprofile}}

Other confidence interval transformations: \code{\link{exp.mcpCI}}
}
\keyword{misc}
