\name{glmobj}
\alias{glmobj}
\title{Build an mcpglm object. For internal
  use only.} 
\description{Constructs a small model object containing several
  estimated parameters.}

\usage{
glmobj(object)
}

\arguments{
  \item{object}{an object of class \code{\link{glm}}}
}

\value{
  A list with model object informations. 
}
\keyword{models}
