\name{hoa-methods}
\docType{methods}
\alias{hoa}
\alias{hoa-methods}
\alias{hoa,mcprofile-method}
\title{Higher Order Approximation of signed root deviance profiles}
\description{
 Applies a saddlepoint approximation to signed root deviance profiles
}
\usage{
hoa(object)
}
\arguments{
  \item{object}{An object of class \code{\linkS4class{mcprofile}}}
}
\value{
  An object of class \code{\linkS4class{mcprofile}}
}
\keyword{methods}
\keyword{models}
