\name{hoa}
\alias{hoa}
\title{Higher order approximation (Test-Version)}
\description{Transforms a signed root deviance profile to a modified
  deviance profile. This is only a test version. Does not use the
  correct Fisher information matrix...}
\usage{
hoa(object)
}
\arguments{
  \item{object}{An object of class mcprofile}
}
\value{An object of class mcprofile with a hoa profile in the srdp slot.}

\author{Daniel Gerhard}

\seealso{\code{\link{mcprofile}}}

\keyword{ misc }
