\name{mcprofile-package}
\alias{mcprofile-package}
\alias{mcprofile}
\docType{package}
\title{Multiple Contrast Profiles}
\description{Calculation of signed root deviance profiles for linear combinations of parameters in a glm. Multiple tests and simultaneous confidence intervals are provided.}
\details{
\tabular{ll}{
Package: \tab mcprofile\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2011-09-09\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Daniel Gerhard

Maintainer: Daniel Gerhard <gerhard@biostat.uni-hannover.de>
}

\seealso{\code{\link{mcpcalc}}}

\keyword{ package }


