\name{mcprofile}
\alias{mcprofile}
\title{General Information on the mcprofile Package}
\description{ Calculates signed root deviance profiles for linear
  combinations of parameters. Simultaneous confidence intervals and
  adjusted p-values are provided. }
\author{Daniel Gerhard <gerhard@biostat.uni-hannover.de>}
\seealso{\code{\link{mcpcalc}}, \code{\link{mcpcalcRatio}}, \code{\link{profile.glm}}, \code{\link{glht}}}
\keyword{package}

