\name{show-methods}
\docType{methods}
\alias{show,mcpconfint-method}
\alias{show,mcprofile-method}
\alias{show,mcprofileRatio-method}
\alias{show,mcptest-method}
\title{ Methods for Function show in Package `mcprofile' }
\description{
 Methods for function \code{show} in Package `mcprofile'
}
\keyword{methods}


