\name{srdp}
\alias{srdp}
\title{Signed root deviance calculation}
\description{Only for internal use.}
\usage{
srdp(cobject, oobject)
}
\arguments{
  \item{cobject}{an mcpglm object with the conditional model estimates}
  \item{oobject}{an mcpglm object with the original model estimates}
}
\keyword{htest}
