\name{transformation-methods}
\docType{methods}
\alias{exp,mcpconfint-method}
\alias{expit}
\alias{expit,mcpconfint-method}

\title{Transformation of confidence limits}
\description{
  Transform the confidence limits by exponentiating or the expit function.
}
\arguments{
  \item{object}{an object of class \code{\linkS4class{mcpconfint}}}
  \item{x}{an object of class numeric}
}
\details{
  Only the confidence limits and point estimates are transformed.
}

\keyword{methods}
\keyword{manip}
\keyword{arith}
