\name{wald-methods}
\docType{methods}
\alias{wald}
\alias{wald-methods}
\alias{wald,mcprofile-method}
\title{Quadratic approximation of signed root deviance profiles}
\description{
 Approximates signed root deviance profiles by a quadratic function.
}
\usage{
wald(object)
}
\arguments{
  \item{object}{An object of class \code{\linkS4class{mcprofile}}}
}
\value{
  An object of class \code{\linkS4class{mcprofile}}
}

\keyword{methods}
\keyword{models}
