\name{wald}
\alias{wald}
\title{Calculate Wald-Profiles}
\description{Transforms a signed root deviance profile of a mcprofile
  object into a wald-profile}
\usage{
wald(object)
}
\arguments{
  \item{object}{An object of class mcprofile}
}
\value{An object of class mcprofile with a wald profile in the srdp slot.}

\author{Daniel Gerhard}

\seealso{\code{\link{mcprofile}}}

\keyword{ misc }
