\name{cta}
\alias{cta}
\docType{data}
\title{Cell transformation assay dataset}
\description{Balb/c 3T3 cells are treated with different concentrations of a carcinogen. Cells treated with a carcinogen do not stop proliferation. Number of foci (cell accumulations) are counted for 10 replicates per concentration level.}
\usage{data(cta)}
\format{
  A data frame with 80 observations on the following 2 variables.
  \describe{
    \item{\code{conc}}{a numeric vector denoting the concentration levels}
    \item{\code{foci}}{a numeric vector with the number of foci}
  }
}
\references{Thomas C (2008): ECVAM data}
\keyword{datasets}
